# openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  version: "1.0.0"
  description: API for Order Service (Phase 1)
servers:
  - url: http://localhost:3000
paths:
  /orders:
    post:
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List orders (placeholder for future implementation)
      responses:
        '501':
          description: Not implemented
  /orders/{id}:
    get:
      summary: Get order by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    CreateOrder:
      type: object
      required: [userId, items, total]
      properties:
        userId:
          type: string
          minLength: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          minItems: 1
        total:
          type: number
          minimum: 0
    OrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 1
    Order:
      type: object
      required: [id, userId, items, total, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total:
          type: number
        status:
          type: string
          enum: [created, reserved, prepared, failed]
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string